Implement token transactions module
